Move RecordHeader back to core package
