+++ caps-spark/src/main/scala/org/opencypher/caps/api/spark/CAPSRecords.scala
@@ -28,11 +28,12 @@ import org.opencypher.caps.api.types._
 import org.opencypher.caps.api.util.PrintOptions
 import org.opencypher.caps.api.value.CypherMap
 import org.opencypher.caps.impl.record.CAPSRecordHeader
+import org.opencypher.caps.impl.record.CAPSRecordHeader._
+import org.opencypher.caps.impl.spark.DfUtils._
 import org.opencypher.caps.impl.spark.convert.{fromSparkType, rowToCypherMap, toSparkType}
 import org.opencypher.caps.impl.spark.exception.Raise
 import org.opencypher.caps.impl.spark.{RecordsPrinter, SparkColumnName}
 import org.opencypher.caps.impl.syntax.RecordHeaderSyntax._
-import org.opencypher.caps.impl.spark.DfUtils._
 
 import scala.annotation.tailrec
 import scala.reflect.runtime.universe.TypeTag
