+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/RowExpansion.scala
@@ -20,7 +20,9 @@ import org.apache.spark.sql.types.StructType
 import org.opencypher.caps.api.expr._
 import org.opencypher.caps.api.record.{ProjectedExpr, RecordHeader, RecordSlot}
 import org.opencypher.caps.api.types.{CTNode, CTRelationship}
+import org.opencypher.caps.impl.record.CAPSRecordHeader._
 import org.opencypher.caps.impl.spark.exception.Raise
+
 
 case class RowExpansion(
   targetHeader: RecordHeader,
