--- caps-spark/src/main/scala/org/opencypher/caps/impl/syntax/AllSyntax.scala
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2016-2017 "Neo4j, Inc." [https://neo4j.com]
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.opencypher.caps.impl.syntax
-
-import org.opencypher.caps.common.syntax.RegisterSyntax
-
-trait AllSyntax
-  extends RegisterSyntax
-  with RecordHeaderSyntax
